refactor: ChessService 클래스 분리
